--- a/staging/src/k8s.io/apimachinery/pkg/util/cache/lruexpirecache_test.go
+++ b/staging/src/k8s.io/apimachinery/pkg/util/cache/lruexpirecache_test.go
@@ -17,6 +17,8 @@ limitations under the License.
 package cache
 
 import (
+	"fmt"
+	"sync"
 	"testing"
 	"time"
 
@@ -45,7 +47,7 @@ func TestSimpleGet(t *testing.T) {
 }
 
 func TestExpiredGet(t *testing.T) {
-	fakeClock := clock.NewFakeClock(time.Now())
+	fakeClock := clock.NewFakeClock(time.Time{})
 	c := NewLRUExpireCacheWithClock(10, fakeClock)
 	c.Add("short-lived", "12345", 1*time.Millisecond)
 	// ensure the entry expired
@@ -65,4 +67,401 @@ func TestLRUOverflow(t *testing.T) {
 	expectEntry(t, c, "elem3", "3")
 	expectEntry(t, c, "elem4", "4")
 	expectEntry(t, c, "elem5", "5")
+}
+
+type joiningTestFramework struct {
+	clock *clock.FakeClock
+	cache *LRUExpireCache
+}
+
+func newJoiningTest(maxEntries int) *joiningTestFramework {
+	j := &joiningTestFramework{
+		clock: clock.NewFakeClock(time.Time{}),
+	}
+	j.cache = NewLRUExpireCacheWithClock(maxEntries, j.clock)
+	return j
+}
+
+type blockedComputeFunc struct {
+	framework *joiningTestFramework
+
+	key   string
+	value string
+
+	// if a race is expected, the winning compute function sets this to true.
+	winner bool
+
+	computeTTL    time.Duration
+	cacheEntryTTL time.Duration
+
+	callCount  int
+	blockUntil chan struct{}
+}
+
+func newBCF(j *joiningTestFramework, k, v string, computeTTL, entryTTL time.Duration) *blockedComputeFunc {
+	return &blockedComputeFunc{
+		framework:     j,
+		key:           k,
+		value:         v,
+		computeTTL:    computeTTL,
+		cacheEntryTTL: entryTTL,
+		blockUntil:    make(chan struct{}),
+	}
+}
+
+func (b *blockedComputeFunc) done() {
+	close(b.blockUntil)
+}
+
+func (b *blockedComputeFunc) compute(notifyOnBegin chan<- struct{}) func(abort <-chan time.Time) (interface{}, time.Duration) {
+	return func(abort <-chan time.Time) (interface{}, time.Duration) {
+		notifyOnBegin <- struct{}{}
+		// fmt.Printf("compute[%v=%v] begin: %v\n", b.key, b.value, b.framework.clock.Now())
+		b.callCount++
+
+		// Make sure that if both channels are waiting when we get
+		// here, the abort channel is selected.
+		select {
+		case <-abort:
+			// fmt.Printf("compute[%v=%v] aborting: %v\n", b.key, b.value, b.framework.clock.Now())
+			return nil, 0
+		default:
+		}
+
+		select {
+		case <-b.blockUntil:
+			// fmt.Printf("compute[%v=%v] unblocked: %v\n", b.key, b.value, b.framework.clock.Now())
+			return b.value, b.cacheEntryTTL
+		case <-abort:
+			// fmt.Printf("compute[%v=%v] aborting: %v\n", b.key, b.value, b.framework.clock.Now())
+			return nil, 0
+		}
+	}
+}
+
+// returns true if a thread was actually unblocked
+func (b *blockedComputeFunc) tryUnblockOne() bool {
+	select {
+	case b.blockUntil <- struct{}{}:
+		return true
+	default:
+	}
+	// This can happen if all compute() funcs have exited, OR if they
+	// haven't gotten far enough to wait on the channel yet -- we have to
+	// count starts before trying unblocks to avoid this..
+	return false
+}
+
+// unblockAtLeast attempts to unblock calls, and stops when it unblocks n
+// calls, or after 100 attempts.
+// If n is zero, make 100 attempts (for verifying no unexpected threads were pending).
+//
+// Returns the number of unblocked calls.
+func (b *blockedComputeFunc) unblockAtLeast(n int) int {
+	got := 0
+	for i := 0; i < n+2; i++ {
+		if b.tryUnblockOne() {
+			got++
+			if got == n {
+				break
+			}
+		}
+	}
+	return got
+}
+
+// checkOne calls GetOrWait.
+// wg is incremented and wg.Done() is called after the GetOrWait call has exited.
+// One signal is sent down notifyOnBegin, either when the computeFunc enters,
+// or immediately when GetOrWait finishes (if the computeFunc wasn't called due
+// to a cache hit). eval() is called on the result of GetOrWait.
+func (b *blockedComputeFunc) checkOne(wg *sync.WaitGroup, notifyOnBegin chan<- struct{}, eval func(interface{}, bool)) {
+	wg.Add(1)
+	ensureStarted := make(chan struct{}, 2) // might have 2 entries written to it, and we don't wish to block.
+	go func() {
+		defer wg.Done()
+		v, ok := b.framework.cache.GetOrWait(b.key, b.compute(ensureStarted), b.computeTTL)
+
+		// we don't know if we'll enter the compute func or not, so ping this just in case.
+		ensureStarted <- struct{}{}
+		// fmt.Println("pinging ensureStarted redundantly")
+		eval(v, ok)
+	}()
+	go func() {
+		<-ensureStarted
+		// fmt.Println("definitely GetOrWait started somehow")
+		notifyOnBegin <- struct{}{}
+	}()
+}
+
+func (b *blockedComputeFunc) expectExactly(t *testing.T) func(interface{}, bool) {
+	return func(v interface{}, ok bool) {
+		if !ok {
+			t.Errorf("unexpected cache miss")
+		}
+		if v != b.value {
+			t.Errorf("unexpected contents: %v", v)
+		}
+	}
+}
+
+func (b *blockedComputeFunc) expectMiss(t *testing.T) func(interface{}, bool) {
+	return func(v interface{}, ok bool) {
+		if ok {
+			t.Errorf("expected cache miss but got %v", v)
+		}
+	}
+}
+
+func (b *blockedComputeFunc) expectRace(t *testing.T) func(interface{}, bool) {
+	return func(v interface{}, ok bool) {
+		if ok && v == b.value {
+			b.winner = true
+		}
+	}
+}
+
+func TestJoinedGet(t *testing.T) {
+	t.Parallel()
+	j := newJoiningTest(10)
+
+	cf := newBCF(j, "joined", "12345", 2*time.Second, 10*time.Second)
+	defer cf.done()
+
+	var wg sync.WaitGroup
+	startCounter := make(chan struct{}, 2)
+
+	cf.checkOne(&wg, startCounter, cf.expectExactly(t))
+	j.clock.Step(500 * time.Millisecond)
+
+	cf.checkOne(&wg, startCounter, cf.expectExactly(t))
+	j.clock.Step(1500 * time.Millisecond)
+
+	<-startCounter
+
+	// The two calls to unblockAtLeast are to notice if a second compute function was started.
+	unblocked := cf.unblockAtLeast(1)
+	<-startCounter
+	unblocked += cf.unblockAtLeast(0)
+
+	if unblocked != 1 {
+		t.Errorf("expected to unblock %v waiting compute() calls, but unblocked %v", 1, unblocked)
+	}
+
+	wg.Wait()
+
+	if cf.callCount != 1 {
+		t.Errorf("saw %v calls", cf.callCount)
+	}
+
+	expectEntry(t, j.cache, "joined", "12345")
+}
+
+func TestJoinedGetTimeout(t *testing.T) {
+	var table = []struct {
+		wait            time.Duration
+		expectTimeout   bool
+		expectUnblocked int
+		expectCalls     int
+	}{
+		{1999 * time.Millisecond, false, 1, 1},
+		{2001 * time.Millisecond, true, 0, 1},
+	}
+	for _, tt := range table {
+		t.Run("", func(t *testing.T) {
+			j := newJoiningTest(10)
+
+			cf := newBCF(j, "joined", "12345", 2*time.Second, 10*time.Second)
+			defer cf.done()
+
+			var wg sync.WaitGroup
+			startCounter := make(chan struct{}, 2)
+
+			if tt.expectTimeout {
+				cf.checkOne(&wg, startCounter, cf.expectMiss(t))
+			} else {
+				cf.checkOne(&wg, startCounter, cf.expectExactly(t))
+			}
+			<-startCounter
+			j.clock.Step(tt.wait)
+
+			unblocked := cf.unblockAtLeast(1)
+			wg.Wait()
+
+			if tt.expectTimeout {
+				expectNotEntry(t, j.cache, "joined")
+			} else {
+				expectEntry(t, j.cache, "joined", "12345")
+			}
+
+			if e, a := tt.expectUnblocked, unblocked; e != a {
+				t.Errorf("expected to unblock %v waiting compute() calls, but unblocked %v", e, a)
+			}
+
+			if e, a := tt.expectCalls, cf.callCount; e != a {
+				t.Errorf("expected %v calls but saw %v calls", e, a)
+			}
+		})
+	}
+}
+
+func TestRegularGetDuringJoinedGet(t *testing.T) {
+	j := newJoiningTest(10)
+
+	cf := newBCF(j, "joined", "12345", 2*time.Second, 10*time.Second)
+	defer cf.done()
+
+	var wg sync.WaitGroup
+	startCounter := make(chan struct{}, 2)
+
+	cf.checkOne(&wg, startCounter, cf.expectExactly(t))
+	<-startCounter
+	j.clock.Step(1800 * time.Millisecond)
+
+	// An entry is in the cache, but a regular get doesn't block on it.
+	expectNotEntry(t, j.cache, "joined")
+
+	unblocked := cf.unblockAtLeast(1)
+	wg.Wait()
+
+	expectEntry(t, j.cache, "joined", "12345")
+
+	if e, a := 1, unblocked; e != a {
+		t.Errorf("expected to unblock %v waiting compute() calls, but unblocked %v", e, a)
+	}
+
+	if e, a := 1, cf.callCount; e != a {
+		t.Errorf("expected %v calls but saw %v calls", e, a)
+	}
+}
+
+func TestEvictDuringJoinedGet(t *testing.T) {
+	j := newJoiningTest(3)
+
+	cf := newBCF(j, "joined", "12345", 2*time.Second, 10*time.Second)
+	defer cf.done()
+
+	var wg sync.WaitGroup
+	startCounter := make(chan struct{}, 2)
+
+	cf.checkOne(&wg, startCounter, cf.expectExactly(t))
+	<-startCounter
+	j.clock.Step(1800 * time.Millisecond)
+
+	j.cache.Add("elem1", "1", 10*time.Hour)
+	j.cache.Add("elem2", "2", 10*time.Hour)
+	j.cache.Add("elem3", "3", 10*time.Hour)
+
+	// Because the first entry has been evicted already, this will cause a
+	// second call and not be joined.
+	cf.checkOne(&wg, startCounter, cf.expectExactly(t))
+	<-startCounter
+
+	unblocked := cf.unblockAtLeast(2)
+	wg.Wait()
+
+	expectEntry(t, j.cache, "joined", "12345")
+
+	if e, a := 2, unblocked; e != a {
+		t.Errorf("expected to unblock %v waiting compute() calls, but unblocked %v", e, a)
+	}
+
+	if e, a := 2, cf.callCount; e != a {
+		t.Errorf("expected %v calls but saw %v calls", e, a)
+	}
+}
+
+func TestJoinedThrashing(t *testing.T) {
+	j := newJoiningTest(20)
+	stop := make(chan struct{})
+	var wg sync.WaitGroup
+	startCounter := make(chan struct{}, 1000)
+	wg.Add(2)
+	go func() {
+		defer wg.Done()
+		for {
+			select {
+			case <-stop:
+				return
+			case <-startCounter:
+			}
+		}
+	}()
+	go func() {
+		defer wg.Done()
+		for {
+			select {
+			case <-stop:
+				return
+			default:
+				j.clock.Step(1700 * time.Millisecond)
+			}
+		}
+	}()
+
+	unblocker := func(cf *blockedComputeFunc) {
+		wg.Add(1)
+		go func() {
+			defer wg.Done()
+			defer cf.done()
+			unblocked := 0
+			for {
+				unblocked += cf.unblockAtLeast(1)
+				select {
+				case <-stop:
+					return
+				default:
+				}
+			}
+			t.Logf("unblocked %v total threads", unblocked)
+		}()
+	}
+
+	for i := 0; i < 10; i++ {
+		wg.Add(3)
+		str := fmt.Sprintf("%v", i)
+		go func() {
+			defer wg.Done()
+			cf := newBCF(j, "race_"+str, "12345", 2*time.Second, 10*time.Second)
+			go unblocker(cf)
+			for {
+				cf.checkOne(&wg, startCounter, cf.expectRace(t))
+				select {
+				case <-stop:
+					return
+				default:
+				}
+			}
+		}()
+		go func() {
+			defer wg.Done()
+			cf := newBCF(j, "race_"+str, "54321", 2*time.Second, 2*time.Second)
+			go unblocker(cf)
+			for {
+				cf.checkOne(&wg, startCounter, cf.expectRace(t))
+				select {
+				case <-stop:
+					return
+				default:
+				}
+			}
+		}()
+		go func() {
+			defer wg.Done()
+			cf := newBCF(j, "norace_"+str, "12345", 2*time.Second, 2*time.Second)
+			go unblocker(cf)
+			for {
+				cf.checkOne(&wg, startCounter, cf.expectRace(t))
+				select {
+				case <-stop:
+					return
+				default:
+				}
+			}
+		}()
+	}
+
+	time.Sleep(5 * time.Second)
+	close(stop)
+	wg.Wait()
 }