--- a/staging/src/k8s.io/apiserver/pkg/authentication/token/cache/cache_simple.go
+++ b/staging/src/k8s.io/apiserver/pkg/authentication/token/cache/cache_simple.go
@@ -40,6 +40,15 @@ func (c *simpleCache) get(key string) (*cacheRecord, bool) {
 	return value, ok
 }
 
+func (c *simpleCache) getOrWait(key string, compute lrucache.ComputeFunc, computationTime time.Duration) (value *cacheRecord, exists bool) {
+	val, ok := c.lru.GetOrWait(key, compute, computationTime)
+	if !ok {
+		return nil, false
+	}
+	value, ok := record.(*cacheRecord)
+	return value, ok
+}
+
 func (c *simpleCache) set(key string, value *cacheRecord, ttl time.Duration) {
 	c.lru.Add(key, value, ttl)
 }