--- a/staging/src/k8s.io/apiserver/pkg/authentication/token/cache/cache_striped.go
+++ b/staging/src/k8s.io/apiserver/pkg/authentication/token/cache/cache_striped.go
@@ -46,6 +46,9 @@ func newStripedCache(stripeCount int, hash hashFunc, newCacheFunc newCacheFunc) 
 func (c *stripedCache) get(key string) (*cacheRecord, bool) {
 	return c.caches[c.hashFunc(key)%c.stripeCount].get(key)
 }
+func (c *simpleCache) getOrWait(key string, compute lrucache.ComputeFunc, computationTime time.Duration) (value *cacheRecord, exists bool) {
+	return c.caches[c.hashFunc(key)%c.stripeCount].getOrWait(key, compute, computationTime)
+}
 func (c *stripedCache) set(key string, value *cacheRecord, ttl time.Duration) {
 	c.caches[c.hashFunc(key)%c.stripeCount].set(key, value, ttl)
 }