/kind bug

**What this PR does / why we need it**:

**Which issue(s) this PR fixes**:
Fixes #83259

```release-note
NONE
```