@lavalamp 2019-10-04
Not sure I'll stick with this approach. Needs more tests and to be wired in to the auth cache, but it's time to go home.