@tedyu 2019-10-07
Looks like staging/src/k8s.io/apimachinery/pkg/util/cache/BUILD missed a dependency.