@lavalamp 2019-10-07
I think I'm happy with the cache now; tests should be pretty comprehensive. Still wiring it up; haven't looked at tests for that yet.