@lavalamp 2019-10-08
It was kind of tricky to write a benchmark that demonstrated the difference, but I did finally manage.

```
$ benchcmp old_cache.txt new_cache.txt
benchmark                                           old ns/op     new ns/op     delta
BenchmarkCachedTokenAuthenticator/toks-100-12       290342        2859          -99.02%
BenchmarkCachedTokenAuthenticator/toks-500-12       342699        3048          -99.11%
BenchmarkCachedTokenAuthenticator/toks-1000-12      287617        3509          -98.78%
BenchmarkCachedTokenAuthenticator/toks-2000-12      368242        4657          -98.74%
BenchmarkCachedTokenAuthenticator/toks-5000-12      490852        190336        -61.22%
BenchmarkCachedTokenAuthenticator/toks-10000-12     408432        247222        -39.47%
```