@mikedanese 2019-10-09
Can we a https://godoc.org/golang.org/x/sync/singleflight per shard instead?