@lavalamp 2019-10-09
Why?

On Tue, Oct 8, 2019, 7:40 PM Mike Danese <notifications@github.com> wrote:

> Can we a https://godoc.org/golang.org/x/sync/singleflight per shard
> instead?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/83519?email_source=notifications&email_token=AAE6BFTKRJBUYR2INKHSCJ3QNVADNA5CNFSM4I5VOTTKYY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOEAWKA5Q#issuecomment-539795574>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAE6BFR5SITUI7KNQHTF2OLQNVADNANCNFSM4I5VOTTA>
> .
>