@mikedanese 2019-10-09
It does what we are trying to do here and the complexity is contained within a library maintained by the Go team. We can protect this section:

https://github.com/kubernetes/kubernetes/blob/6d97d36fbb5a52e9e7520fb326283a53ba1570a4/staging/src/k8s.io/apiserver/pkg/authentication/token/cache/cached_token_authenticator.go#L78-L88

With a singleflight.Group and get the same result.