@lavalamp 2019-10-09
That's not really an improvement, e.g. what happens if a call never
returns? As written, the cache bounds the number of callers who get stuck
in that situation.

I'd consider that more carefully if it were actually in the standard
library, the x libraries haven't been uniformly high quality in my
experience, maybe they're better now, but I'm not inclined to switch. Using
code we don't have to own isn't automatically better...

Anyway the cache as it exists in the code encourages this bad usage
pattern, I either don't want the cache to be in a library I'm on the hook
for at all, or I want it to give you the option of not using the included
foot gun.

On Tue, Oct 8, 2019, 7:56 PM Mike Danese <notifications@github.com> wrote:

> It does what we are trying to do here and the complexity is contained
> within a library maintained by the Go team. We can protect this section:
>
>
> https://github.com/kubernetes/kubernetes/blob/6d97d36fbb5a52e9e7520fb326283a53ba1570a4/staging/src/k8s.io/apiserver/pkg/authentication/token/cache/cached_token_authenticator.go#L78-L88
>
> With a singleflight.Group and get the same result.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/83519?email_source=notifications&email_token=AAE6BFVTIECMLD6NVODFX7LQNVB4TA5CNFSM4I5VOTTKYY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOEAWK5GI#issuecomment-539799193>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAE6BFS2YISZMFKSMLOEOOTQNVB4TANCNFSM4I5VOTTA>
> .
>