@mikedanese 2019-10-09
Yup, there's always a tradeoff.

> I'd consider that more carefully if it were actually in the standard library,

The x/sync/singleflight package is a mirror of src/internal/sync/singleflight. It's used to share lookups in the Go DNS resolver:

https://github.com/golang/go/blob/fac3b5d05ecf31a2491949cc905312a34e272ae8/src/net/lookup.go#L149-L152

So we're actually already using it in apiserver, controller-manager, kubectl, etc compile with CGO_ENABLED=0... whenever we do a DNS lookup.

> what happens if a call never returns?

https://godoc.org/golang.org/x/sync/singleflight#DoChan returns a result chan that a caller can select on with a ctx.Done(). https://godoc.org/golang.org/x/sync/singleflight#Forget allows a caller to abandon a stuck lookup. src/internal/singleflight also has ForgetUnshared which is used in the resolver for this situation:

https://github.com/golang/go/blob/2197321db1dd997165c0091ba2bcb3b6be7633d0/src/net/lookup.go#L277-L284