@lavalamp 2019-10-09
I personally still think it is better to fix the cache. If you can point to a concrete problem in my implementation... ? But right now it just feels like more work to switch to that for no clear purpose.

Fixing it in the cache means only one lock to contend on.

I do want to modify my interface slightly, I think I can use context better. (than either my current code or the inflight one)