@mikedanese 2019-10-09
> If you can point to a concrete problem in my implementation... ?

I can give the change a thorough review, but assuming that all issues with software can be fixed, this devolves to the arguments for and against using external dependencies. This code introduced in this change is fairly complex change that we'll have to maintain and singlfight seems like a solid alternative that fits what we need here.

Other than that, this change is currently coupled with the LRUExpiringCache. I added some reason in https://github.com/kubernetes/kubernetes/issues/83375#issuecomment-540243511 why we might want to move off LRU for the token cache. Decoupling the current implementation is also something that we could do without introducing an external dependency.