@mikedanese 2019-10-09
> hm, actually it's hard to improve on my interface since context isn't testable (doesn't support injecting a clock for testing).

Not sure I follow this.