@lavalamp 2019-10-11
I have unpushed changes that address some of those comments, but I would be even happier to delete the LRU cache completely as you mentioned elsewhere, I don't actually trust the thing it depends on.

I will send this benchmark in a separate PR and hold on to this one for a bit.